chore: split up newrelic model.
