+++ conftool/model/api_endpoint.py
@@ -1,0 +1,6 @@
+from .enum import ConfigEnum
+
+class ApiEndpointModel(ConfigEnum):
+   US = "US"
+   EU = "EU"
+   FEDRAMP = "FEDRAMP"
