+++ conftool/model/data_format.py
@@ -1,0 +1,5 @@
+from .enum import ConfigEnum
+
+class DataFormatModel(ConfigEnum):
+   EVENTS = "events"
+   LOGS = "logs"
